Compute the words prefixes at the end of an update
